:art: refactor : 필요없는DTO 삭제
