--- src/main/java/lotto/step2/dto/LottoNumberDTO.java
@@ -1,13 +1,0 @@
-package lotto.step2.dto;
-
-public class LottoNumberDTO {
-    private final int lottoNumber;
-    
-    public LottoNumberDTO(int lottoNumber) {
-        this.lottoNumber = lottoNumber;
-    }
-    
-    public int getLottoNumber() {
-        return lottoNumber;
-    }
-}
